+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/implementation/models/CreateCallFailed.java
@@ -5,47 +5,52 @@
 package com.azure.communication.callautomation.implementation.models;
 
 import com.azure.core.annotation.Immutable;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
 
-/** The CreateCallFailed event. */
+/**
+ * The CreateCallFailed event.
+ */
 @Immutable
-public final class CreateCallFailed {
+public final class CreateCallFailed implements JsonSerializable<CreateCallFailed> {
     /*
-     * Used by customers when calling mid-call actions to correlate the request
-     * to the response event.
+     * Used by customers when calling mid-call actions to correlate the request to the response event.
      */
-    @JsonProperty(value = "operationContext", access = JsonProperty.Access.WRITE_ONLY)
     private String operationContext;
 
     /*
      * Contains the resulting SIP code, sub-code and message.
      */
-    @JsonProperty(value = "resultInformation", access = JsonProperty.Access.WRITE_ONLY)
     private ResultInformation resultInformation;
 
     /*
      * Call connection ID.
      */
-    @JsonProperty(value = "callConnectionId", access = JsonProperty.Access.WRITE_ONLY)
     private String callConnectionId;
 
     /*
      * Server call ID.
      */
-    @JsonProperty(value = "serverCallId", access = JsonProperty.Access.WRITE_ONLY)
     private String serverCallId;
 
     /*
-     * Correlation ID for event to call correlation. Also called ChainId for
-     * skype chain ID.
+     * Correlation ID for event to call correlation. Also called ChainId for skype chain ID.
      */
-    @JsonProperty(value = "correlationId", access = JsonProperty.Access.WRITE_ONLY)
     private String correlationId;
+
+    /**
+     * Creates an instance of CreateCallFailed class.
+     */
+    public CreateCallFailed() {
+    }
 
     /**
      * Get the operationContext property: Used by customers when calling mid-call actions to correlate the request to
      * the response event.
-     *
+     * 
      * @return the operationContext value.
      */
     public String getOperationContext() {
@@ -54,7 +59,7 @@ public final class CreateCallFailed {
 
     /**
      * Get the resultInformation property: Contains the resulting SIP code, sub-code and message.
-     *
+     * 
      * @return the resultInformation value.
      */
     public ResultInformation getResultInformation() {
@@ -63,7 +68,7 @@ public final class CreateCallFailed {
 
     /**
      * Get the callConnectionId property: Call connection ID.
-     *
+     * 
      * @return the callConnectionId value.
      */
     public String getCallConnectionId() {
@@ -72,7 +77,7 @@ public final class CreateCallFailed {
 
     /**
      * Get the serverCallId property: Server call ID.
-     *
+     * 
      * @return the serverCallId value.
      */
     public String getServerCallId() {
@@ -82,10 +87,53 @@ public final class CreateCallFailed {
     /**
      * Get the correlationId property: Correlation ID for event to call correlation. Also called ChainId for skype chain
      * ID.
-     *
+     * 
      * @return the correlationId value.
      */
     public String getCorrelationId() {
         return this.correlationId;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of CreateCallFailed from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of CreateCallFailed if the JsonReader was pointing to an instance of it, or null if it was
+     * pointing to JSON null.
+     * @throws IOException If an error occurs while reading the CreateCallFailed.
+     */
+    public static CreateCallFailed fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            CreateCallFailed deserializedCreateCallFailed = new CreateCallFailed();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("operationContext".equals(fieldName)) {
+                    deserializedCreateCallFailed.operationContext = reader.getString();
+                } else if ("resultInformation".equals(fieldName)) {
+                    deserializedCreateCallFailed.resultInformation = ResultInformation.fromJson(reader);
+                } else if ("callConnectionId".equals(fieldName)) {
+                    deserializedCreateCallFailed.callConnectionId = reader.getString();
+                } else if ("serverCallId".equals(fieldName)) {
+                    deserializedCreateCallFailed.serverCallId = reader.getString();
+                } else if ("correlationId".equals(fieldName)) {
+                    deserializedCreateCallFailed.correlationId = reader.getString();
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedCreateCallFailed;
+        });
     }
 }