+++ sdk/communication/azure-communication-callautomation/src/main/java/com/azure/communication/callautomation/implementation/models/TranscriptionConfigurationInternal.java
@@ -5,39 +5,46 @@
 package com.azure.communication.callautomation.implementation.models;
 
 import com.azure.core.annotation.Fluent;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.azure.json.JsonReader;
+import com.azure.json.JsonSerializable;
+import com.azure.json.JsonToken;
+import com.azure.json.JsonWriter;
+import java.io.IOException;
 
-/** Configuration of live transcription. */
+/**
+ * Configuration of live transcription.
+ */
 @Fluent
-public final class TranscriptionConfigurationInternal {
+public final class TranscriptionConfigurationInternal implements JsonSerializable<TranscriptionConfigurationInternal> {
     /*
      * Transport URL for live transcription
      */
-    @JsonProperty(value = "transportUrl", required = true)
     private String transportUrl;
 
     /*
      * The type of transport to be used for live transcription, eg. Websocket
      */
-    @JsonProperty(value = "transportType", required = true)
     private TranscriptionTransportTypeInternal transportType;
 
     /*
      * Defines the locale for the data e.g en-CA, en-AU
      */
-    @JsonProperty(value = "locale", required = true)
     private String locale;
 
     /*
-     * Determines if the transcription should be started immediately after call
-     * is answered or not.
+     * Determines if the transcription should be started immediately after call is answered or not.
      */
-    @JsonProperty(value = "startTranscription", required = true)
     private boolean startTranscription;
 
     /**
+     * Creates an instance of TranscriptionConfigurationInternal class.
+     */
+    public TranscriptionConfigurationInternal() {
+    }
+
+    /**
      * Get the transportUrl property: Transport URL for live transcription.
-     *
+     * 
      * @return the transportUrl value.
      */
     public String getTransportUrl() {
@@ -46,7 +53,7 @@ public final class TranscriptionConfigurationInternal {
 
     /**
      * Set the transportUrl property: Transport URL for live transcription.
-     *
+     * 
      * @param transportUrl the transportUrl value to set.
      * @return the TranscriptionConfigurationInternal object itself.
      */
@@ -57,7 +64,7 @@ public final class TranscriptionConfigurationInternal {
 
     /**
      * Get the transportType property: The type of transport to be used for live transcription, eg. Websocket.
-     *
+     * 
      * @return the transportType value.
      */
     public TranscriptionTransportTypeInternal getTransportType() {
@@ -66,7 +73,7 @@ public final class TranscriptionConfigurationInternal {
 
     /**
      * Set the transportType property: The type of transport to be used for live transcription, eg. Websocket.
-     *
+     * 
      * @param transportType the transportType value to set.
      * @return the TranscriptionConfigurationInternal object itself.
      */
@@ -77,7 +84,7 @@ public final class TranscriptionConfigurationInternal {
 
     /**
      * Get the locale property: Defines the locale for the data e.g en-CA, en-AU.
-     *
+     * 
      * @return the locale value.
      */
     public String getLocale() {
@@ -86,7 +93,7 @@ public final class TranscriptionConfigurationInternal {
 
     /**
      * Set the locale property: Defines the locale for the data e.g en-CA, en-AU.
-     *
+     * 
      * @param locale the locale value to set.
      * @return the TranscriptionConfigurationInternal object itself.
      */
@@ -98,7 +105,7 @@ public final class TranscriptionConfigurationInternal {
     /**
      * Get the startTranscription property: Determines if the transcription should be started immediately after call is
      * answered or not.
-     *
+     * 
      * @return the startTranscription value.
      */
     public boolean isStartTranscription() {
@@ -108,12 +115,60 @@ public final class TranscriptionConfigurationInternal {
     /**
      * Set the startTranscription property: Determines if the transcription should be started immediately after call is
      * answered or not.
-     *
+     * 
      * @param startTranscription the startTranscription value to set.
      * @return the TranscriptionConfigurationInternal object itself.
      */
     public TranscriptionConfigurationInternal setStartTranscription(boolean startTranscription) {
         this.startTranscription = startTranscription;
         return this;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public JsonWriter toJson(JsonWriter jsonWriter) throws IOException {
+        jsonWriter.writeStartObject();
+        jsonWriter.writeStringField("transportUrl", this.transportUrl);
+        jsonWriter.writeStringField("transportType", this.transportType == null ? null : this.transportType.toString());
+        jsonWriter.writeStringField("locale", this.locale);
+        jsonWriter.writeBooleanField("startTranscription", this.startTranscription);
+        return jsonWriter.writeEndObject();
+    }
+
+    /**
+     * Reads an instance of TranscriptionConfigurationInternal from the JsonReader.
+     * 
+     * @param jsonReader The JsonReader being read.
+     * @return An instance of TranscriptionConfigurationInternal if the JsonReader was pointing to an instance of it, or
+     * null if it was pointing to JSON null.
+     * @throws IllegalStateException If the deserialized JSON object was missing any required properties.
+     * @throws IOException If an error occurs while reading the TranscriptionConfigurationInternal.
+     */
+    public static TranscriptionConfigurationInternal fromJson(JsonReader jsonReader) throws IOException {
+        return jsonReader.readObject(reader -> {
+            TranscriptionConfigurationInternal deserializedTranscriptionConfigurationInternal
+                = new TranscriptionConfigurationInternal();
+            while (reader.nextToken() != JsonToken.END_OBJECT) {
+                String fieldName = reader.getFieldName();
+                reader.nextToken();
+
+                if ("transportUrl".equals(fieldName)) {
+                    deserializedTranscriptionConfigurationInternal.transportUrl = reader.getString();
+                } else if ("transportType".equals(fieldName)) {
+                    deserializedTranscriptionConfigurationInternal.transportType
+                        = TranscriptionTransportTypeInternal.fromString(reader.getString());
+                } else if ("locale".equals(fieldName)) {
+                    deserializedTranscriptionConfigurationInternal.locale = reader.getString();
+                } else if ("startTranscription".equals(fieldName)) {
+                    deserializedTranscriptionConfigurationInternal.startTranscription = reader.getBoolean();
+                } else {
+                    reader.skipChildren();
+                }
+            }
+
+            return deserializedTranscriptionConfigurationInternal;
+        });
     }
 }